cuizine-api/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # Changed from psycopg2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jwt==1.3.1

# External Services
openai==1.12.0
requests==2.31.0

# Upstash instead of Redis
upstash-redis==1.0.0

# Monitoring & Logging
sentry-sdk==1.40.0
slack-sdk==3.27.0

# Utils
python-dateutil==2.8.2